aylien_app/aylien_requirements.txt
fastapi
uvicorn
aylien-news-api
python-dotenv
elasticsearch
openai
gradio
langchain
tiktoken
streamlit